requirements.txt
-e .

pymongo[srv]==4.11.3
pandas==2.2.3
ensure==1.0.4
python-box==7.3.2
PyYAML==7.3.2
dvc
scipy==1.15.2
joblib
scikit-learn